refactor: consolidate ticket schemas
